chore: update models

+++ models/model.category.go
@@ -2,6 +2,6 @@ package models
 
 type Category struct {
 	CategoryId	string	`json:"CategoryId"`
-	Color				int16		`json:"Color"`
+	Color				int64		`json:"Color"`
 	Label				string	`json:"Label"`
 }
