+++ models/model.wallet.go
@@ -2,6 +2,6 @@ package models
 
 type Wallet struct {
 	WalletId	string	`json:"WalletId"`
-	Color			int16		`json:"Color"`
+	Color			int64		`json:"Color"`
 	Label			string	`json:"Label"`
 }
